# src/main/java/br/com/serviceflow/api/dto/auth/LoginRequest.java
package br.com.serviceflow.api.dto.auth;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record LoginRequest(
        @NotBlank @Email String email,
        @NotBlank @Size(max = 72) String password) {}